server/requirements.txt
requests
pygithub
langgraph
langchain-core
langchain-community
langchain-openai
langchain-groq
langchain-ollama
langchain-anthropic
GitPython